// src/app/train/model.interface.ts
export interface IModel extends IAsset {
  status: IModel.Status;
  dataType?: string;
  tag?: string;
  size?: number;
  filePath?: string;
  experimentId?: TObjectId;
}
// ...
export namespace IModel {
  export enum Status {
    ACTIVE = 'ACTIVE',
    IMPORTING = 'IMPORTING',
    ERROR = 'ERROR',
    CANCELLED = 'CANCELLED',
  }
}